move test to package 'routers'
