+++ service/javadsl/integration-tests/src/test/java/com/lightbend/lagom/it/routers/AdditionalRoutersServiceImpl.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <https://www.lightbend.com>
  */
 
-package com.lightbend.lagom.it.mocks;
+package com.lightbend.lagom.it.routers;
 
 public class AdditionalRoutersServiceImpl implements AdditionalRoutersService {
 }
